+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/New-SESReceiptFilter-Cmdlet.cs
@@ -65,7 +65,7 @@ namespace Amazon.PowerShell.Cmdlets.SES
         #region Parameter Filter_Name
         /// <summary>
         /// <para>
-        /// <para>The name of the IP address filter. The name must:</para><ul><li><para>Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_),
+        /// <para>The name of the IP address filter. The name must:</para><ul><li><para>This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_),
         /// or dashes (-).</para></li><li><para>Start and end with a letter or number.</para></li><li><para>Contain less than 64 characters.</para></li></ul>
         /// </para>
         /// </summary>